requirements.txt
requests
pydantic
pytest
pandas
plotly
psycopg2-binary
fastapi
textual
uvicorn
python-dotenv
